+++ example/app.vue
@@ -1,11 +1,12 @@
 <template>
   <dynamic-form ref="refModules" v-model="data.formValue" :fields="data.fields"/>
   <el-button @click="onClick">tagger ElFormItem</el-button>
+  <dynamic-modules field="div" :children="data.children"/>
 </template>
 
 <script lang="ts">export default { name: 'App' };</script>
 <script lang="ts" setup>
-import { DynamicForm } from 'dynamic-module';
+import { DynamicForm, DynamicModules } from 'dynamic-module';
 import { reactive, ref } from 'vue';
 
 const refModules = ref();
@@ -21,7 +22,14 @@ const data = reactive({
   fields: [
     { field: 'el-form-item', label: '姓名', required: true, children: [{ field: 'el-input', prop: 'form.name', label: '姓名' }] },
     { col: { span: 12 }, field: 'el-form-item', label: '姓名', required: true, children: [{ field: 'el-input', prop: 'form.name', label: '姓名' }] },
-    { col: { span: 12 }, field: 'el-form-item', label: '姓名', required: true, children: [{ field: 'el-input', prop: 'form.name', label: '姓名' }] },
+    { col: { span: 12 }, field: 'el-form-item', label: '姓名', required: true, style: { paddingLeft: '10px' }, children: [{ field: 'el-input', prop: 'form.name', label: '姓名' }] },
   ],
+  children: [{
+    field: 'el-table', data: [{ name: 'ReFunny', age: '18', address: '中国' }], children: [
+      { field: 'el-table-column', prop: 'name', label: '姓名' },
+      { field: 'el-table-column', prop: 'age', label: '年龄' },
+      { field: 'el-table-column', prop: 'address', label: '地址' },
+    ],
+  }],
 });
 </script>