+++ src/core/interface/index.ts
@@ -3,6 +3,7 @@ export interface BaseField {
   field: string;
   prop?: string;
   show?: boolean;
+  slotData?: Record<string, any>;
   children?: (BaseField | any)[];
   slots?: Record<string, (BaseField | any)[]>;
 }