+++ src/core/src/index.vue
@@ -1,9 +1,9 @@
 <template>
-  <component :is="component" v-if="props.show" ref="refCore" v-bind="$attrs" v-model="values">
+  <component :is="component" v-if="props.show" ref="refCore" v-bind="Object.assign({ prop: props.prop }, $attrs)" v-model="values">
     <template v-for="(children, key) in slots" v-slot:[key]="slotData" :key="key">
       <dynamic-models v-for="(child, index) in children" :slotData="slotData" v-bind="child" :key="`${key}-${index}`" :ref="onAddRefs(`${key}-${index}`)"/>
     </template>
-    <template v-slot="slotData">
+    <template v-slot="slotData" v-if="(slots.default || []).length">
       <dynamic-models v-for="(item, key) in slots.default" :slotData="slotData" v-bind="item" :key="`default-${key}`" :ref="onAddRefs(`default-${key}`)"/>
     </template>
   </component>
@@ -21,8 +21,9 @@ const props = withDefaults(defineProps<BaseField>(), { show: () => true, childre
 const models = inject<Ref<{ [key: string]: any }>>('dynamic-modules', ref({}));
 const logger = inject<LoggerMessage | Console>('logger-message', console);
 
-const component = computed(() => models.value[props.field] || h(resolveComponent(props.field)));
-if (!component.value) logger.warn(`%s: ${ props.field } is not found`, 'Dynamic Modules');
+const defaultType = ['div', 'span', 'p'];
+const component = computed(() => models.value[props.field] || h(!defaultType.includes(props.field) ? resolveComponent(props.field) : props.field));
+if (!component.value) logger.warn(`%s Failed to resolve component: %s`, 'Dynamic Modules', props.field);
 
 const slots = computed(() => {
   const target = props.slots || {};